backend/requirements.txt
# Django core
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Async / WebSocket
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Celery + Redis
celery==5.4.0
redis==5.0.8
django-redis==5.4.0

# Database
mysqlclient==2.2.4

# Scraping
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.22.0
webdriver-manager==4.0.2

# AI / RAG
anthropic==0.34.2
sentence-transformers==3.1.1
chromadb==0.5.15
rank-bm25==0.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
